Added sweetalert, update&delete

+++ src/app/budget/budget-faculte/budget-faculte.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {BudgetService} from '../../controller/service/budget.service';
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-budget-faculte',
   templateUrl: './budget-faculte.component.html',
@@ -36,14 +36,33 @@ export class BudgetFaculteComponent implements OnInit {
     return this.budgetService.budgetFacultes;
   }
   public bfinfo() {
+    if (this.budgetFaculteVo!=null) {
     this.budgetService.findAllByAnnee();
+    }
   }
 
   public delete(annee: number) {
-    if (annee!=null){
-      this.budgetService.deleteBudgetFaculte(annee).subscribe();
-      this.budgetService.refreshAllFromBf();
-    }
+    Swal({
+      title: 'Etes-vous sure?',
+      text: "Vous ne pouvez pas revenir en arrière!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Oui, supprimer!'
+    }).then((result) => {
+      if (result.value) {
+        if (annee!=null){
+          this.budgetService.deleteBudgetFaculte(annee).subscribe();
+          this.budgetService.refreshAllFromBf();
+        }
+        Swal(
+          'Supprimmé!',
+          'Vos données ont été supprimés.',
+          'success'
+        );
+      }
+    });
   }
 
   public get date(){
